Chat Panel is closable
[+] The Chat panel is now can be closed and reopened
[+] The Chat panel is disabled in the main menu, and it will be first enabled when joined a lobby

+++ Assets/Server/Scripts/ServerSender.cs
@@ -97,7 +97,7 @@ namespace ServerSide
             NetworkManager.Instance.Server.SendToAll(response);
         }
 
-        public static void SendChatMessage(int clientID, string text)
+        public static void SendChatMessage(ushort clientID, string text)
         {
             Message response = Message.Create(MessageSendMode.reliable, ServerToClientPacket.SendMessage);
             response.Add(text);
